+++ platformio/builder/tools/platformio.py
@@ -22,7 +22,7 @@ from os.path import basename, dirname, isdir, join, realpath
 
 from SCons import Builder, Util
 from SCons.Script import (COMMAND_LINE_TARGETS, AlwaysBuild,
-                          DefaultEnvironment, SConscript)
+                          DefaultEnvironment, Export, SConscript)
 
 from platformio.util import glob_escape, pioversion_to_intstr
 
@@ -39,6 +39,35 @@ def scons_patched_match_splitext(path, suffixes=None):
     if suffixes and tokens[1] and tokens[1] in suffixes:
         return (path, tokens[1])
     return tokens
+
+
+def _build_project_deps(env):
+    deps = env.BuildProjectLibraries()
+    # prepend dependent libs before built-in
+    env.Prepend(LIBS=deps['LIBS'])
+
+    if "__test" in COMMAND_LINE_TARGETS:
+        env.ProcessTest()
+        projenv = env.Clone()
+        projenv.BuildSources("$BUILDTEST_DIR", "$PROJECTTEST_DIR",
+                             "$PIOTEST_SRC_FILTER")
+    else:
+        projenv = env.Clone()
+        projenv.BuildSources("$BUILDSRC_DIR", "$PROJECTSRC_DIR",
+                             env.get("SRC_FILTER"))
+
+    # CPPPATH from dependencies
+    projenv.PrependUnique(CPPPATH=deps['CPPPATH'])
+    # extra build flags from `platformio.ini`
+    projenv.ProcessFlags(env.get("SRC_BUILD_FLAGS"))
+
+    if not env.get("PIOBUILDFILES") and not COMMAND_LINE_TARGETS:
+        sys.stderr.write(
+            "Error: Nothing to build. Please put your source code files "
+            "to '%s' folder\n" % env.subst("$PROJECTSRC_DIR"))
+        env.Exit(1)
+
+    Export("projenv")
 
 
 def BuildProgram(env):
@@ -62,6 +91,7 @@ def BuildProgram(env):
     # process extra flags from board
     if "BOARD" in env and "build.extra_flags" in env.BoardConfig():
         env.ProcessFlags(env.BoardConfig().get("build.extra_flags"))
+
     # apply user flags
     env.ProcessFlags(env.get("BUILD_FLAGS"))
 
@@ -74,36 +104,18 @@ def BuildProgram(env):
     # remove specified flags
     env.ProcessUnFlags(env.get("BUILD_UNFLAGS"))
 
-    # build dependent libs; place them before built-in libs
-    env.Prepend(LIBS=env.BuildProjectLibraries())
+    # build project with dependencies
+    _build_project_deps(env)
 
-    # append specified LD_SCRIPT
-    if ("LDSCRIPT_PATH" in env
+    # append into the beginning a main LD script
+    if (env.get("LDSCRIPT_PATH")
             and not any("-Wl,-T" in f for f in env['LINKFLAGS'])):
-        env.Append(LINKFLAGS=['-Wl,-T"$LDSCRIPT_PATH"'])
+        env.Prepend(LINKFLAGS=["-T", "$LDSCRIPT_PATH"])
 
     # enable "cyclic reference" for linker
     if env.get("LIBS") and env.GetCompilerType() == "gcc":
         env.Prepend(_LIBFLAGS="-Wl,--start-group ")
         env.Append(_LIBFLAGS=" -Wl,--end-group")
-
-    # Handle SRC_BUILD_FLAGS
-    env.ProcessFlags(env.get("SRC_BUILD_FLAGS"))
-
-    if "__test" in COMMAND_LINE_TARGETS:
-        env.Append(PIOBUILDFILES=env.ProcessTest())
-    else:
-        env.Append(
-            PIOBUILDFILES=env.CollectBuildFiles(
-                "$BUILDSRC_DIR",
-                "$PROJECTSRC_DIR",
-                src_filter=env.get("SRC_FILTER")))
-
-    if not env['PIOBUILDFILES'] and not COMMAND_LINE_TARGETS:
-        sys.stderr.write(
-            "Error: Nothing to build. Please put your source code files "
-            "to '%s' folder\n" % env.subst("$PROJECTSRC_DIR"))
-        env.Exit(1)
 
     program = env.Program(
         join("$BUILD_DIR", env.subst("$PROGNAME")), env['PIOBUILDFILES'])
@@ -276,7 +288,7 @@ def BuildFrameworks(env, frameworks):
             env.ConvertInoToCpp()
 
         if f in board_frameworks:
-            SConscript(env.GetFrameworkScript(f))
+            SConscript(env.GetFrameworkScript(f), exports="env")
         else:
             sys.stderr.write(
                 "Error: This board doesn't support %s framework!\n" % f)
@@ -290,8 +302,9 @@ def BuildLibrary(env, variant_dir, src_dir, src_filter=None):
 
 
 def BuildSources(env, variant_dir, src_dir, src_filter=None):
+    nodes = env.CollectBuildFiles(variant_dir, src_dir, src_filter)
     DefaultEnvironment().Append(
-        PIOBUILDFILES=env.CollectBuildFiles(variant_dir, src_dir, src_filter))
+        PIOBUILDFILES=[env.Object(node) for node in nodes])
 
 
 def exists(_):